template_site/urls/10-local.py
from django.conf.urls.defaults import patterns, url
from django.conf import settings

urlpatterns += patterns('',
    url(r'^uploads/(?P<path>.*)$', 'django.views.static.serve',
        {'document_root': settings.MEDIA_ROOT, 'show_indexes': True}),
    url(r'^static/(?P<path>.*)$', 'django.views.static.serve',
        {'document_root': settings.STATIC_ROOT, 'show_indexes': True}),
)
