template_site/requirements.txt
south
django_compressor
-e git+https://github.com/avoine/Lesscpy.git#egg=lesscpy
